requirements.txt
asgiref==3.3.1
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==4.0.0
ciso8601==2.1.3
dj-database-url==0.5.0
Django==3.1.5
django-bootstrap4==2.3.1
django-heroku==0.3.1
gunicorn==20.0.4
idna==2.10
importlib-metadata==2.1.1
psycopg2==2.8.6
python-dotenv==0.15.0
pytz==2020.5
requests==2.25.1
soupsieve==2.1
sqlparse==0.4.1
urllib3==1.26.3
whitenoise==5.2.0
zipp==3.4.0
